changed cow poetry into ints to avoid TLE on XJOI, it works

+++ codes/XJOI/XJOI_1625_1.cpp
@@ -18,16 +18,16 @@
 using namespace std;
 
 const long long mod = 1e9 + 7;
-long long dp[max_v][max_v];
-long long c[max_v], s[max_v], groups[120];
-long long n, m, k;
+int  dp[max_v][max_v];
+int c[max_v], s[max_v], groups[120];
+int n, m, k;
 
-long long bin_pow(long long a, long long b){
+int bin_pow(int a, int b){
   //a^b % c;
   long long ans = 1ll;
   long long ind = (long long)a;
-  for(int i = 0; (long long)pow_2(i) <= b; i++){
-    if(b & (long long)pow_2(i)){
+  for(int i = 0; pow_2(i) <= b; i++){
+    if(b & pow_2(i)){
       ans *= ind;
       ans %= mod;
     }
@@ -41,9 +41,9 @@ long long bin_pow(long long a, long long b){
 int main(){
   //freopen("poetry.in", "r", stdin);
   //freopen("poetry.out", "w+", stdout);
-	scanf("%lld%lld%lld", &n, &m, &k);
+	scanf("%d%d%d", &n, &m, &k);
   for(int i = 0; i<n; i++){
-    scanf("%lld%lld", &s[i], &c[i]);
+    scanf("%d%d", &s[i], &c[i]);
   }
   for(int i = 0; i<m; i++){
     static char str[10];
@@ -51,9 +51,9 @@ int main(){
     groups[str[0]]++;
   }
   
-  dp[0][0] = 1ll;
+  dp[0][0] = 1;
   for(int i = 0; i<=k; i++){
-    long long t = 0ll;
+    int t = 0;
     for(int j = 0; j<=n; j++) t = (t + dp[i][j]) % mod;
     
     for(int j = 0; j<n; j++){
@@ -71,11 +71,11 @@ int main(){
     long long temp = 0ll;
     if(groups[i] == 0) cont;
     for(int j = 1; j<=n; j++){ //over every group, if none in group, dp would be 0 so no effect
-      (temp += bin_pow(dp[k][j], groups[i])) %= mod;
+      (temp += (long long)bin_pow(dp[k][j], groups[i])) %= (long long)mod;
     }
     //printf("for %d there are %lld ways\n", i, temp);
     ans *= temp;
-    ans %= mod;
+    ans %= (long long)mod;
   }
 
   printf("%lld\n", ans);
